feat: mock responses outside cypress
